+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.0
+
+* Add more per-request overrides to give more dynamic possibilities for ways to pass the token to AuthPASETO
+
+## 0.5.5
+
+* Add request json as possible token location
+
 ## 0.5.4
 
 * Fix incorrect way to get seconds from datetime.timedelta instance